+++ mocap_popy/scripts/pipeline_runner/batch_runner.py
@@ -2,7 +2,29 @@
     Batch Runner
     ============
 
-    This script runs the NUSHU pipeline runner on a batch of trials. It is intended to be run as a standalone script.
+    This script runs the pipeline runner on a batch of trials. It is intended to be run as a standalone script.
+    The script extracts pipelines from a configuration file. By default, the script looks for a configuration file
+    in the `config` directory of this script's parent.
+
+    The export flag of the pipeline runner is set to true by default. Pipelines should include a pipeline step which
+    exports the trial to C3D, or otherwise save the trial (not recommended).
+
+    Usage:
+    ------
+
+    python batch_runner.py -cn <config_name> -pp <project_path> --exclude <trials_to_exclude>
+
+    Example: 
+        python batch_runner.py -v -l -cn "nushu_pipeline_series.json" -pp "D:\HPL\pipeline_test_2" --include "20241107T104520Z_step-forward-r,20241107T104613Z_step-forward-l"
+
+    Options:
+    --------
+    Run 'python batch_runner.py -h' for options.
+
+    Returns:
+    --------
+    0 if successful, -1 if failed.
+
 
     @author: C. McCarthy
 """
@@ -15,7 +37,7 @@ import time
 
 from mocap_popy.config import logger, directory
 from mocap_popy.utils import dist_utils
-import mocap_popy.scripts.nushu_pipeline_runner.nushu_pipeline_runner as npr
+import mocap_popy.scripts.pipeline_runner.pipeline_runner as npr
 
 LOGGER = logging.getLogger("BatchRunner")
 
@@ -41,22 +63,35 @@ def configure_parser():
     )
 
     parser.add_argument(
+        "-cn",
+        "--config_name",
+        type=str,
+        required=True,
+        help="Name of the pipeline configuration file.",
+    )
+    parser.add_argument(
         "-pp",
         "--project_path",
         type=str,
         required=True,
         help="Path to the project directory",
     )
-
+    parser.add_argument(
+        "--include",
+        type=str,
+        default="",
+        help="Comma-separated list of trials to include",
+    )
     parser.add_argument(
         "--exclude",
         type=str,
+        default="",
         help="Comma-separated list of trials to exclude",
     )
 
     parser.add_argument(
         "-sn",
-        "--subject-name",
+        "--subject_name",
         type=str,
         help="Name of the subject",
     )
@@ -78,6 +113,8 @@ def main():
     if args.verbose:
         LOGGER.setLevel(logging.DEBUG)
 
+    config_path = npr.validate_config_name(args.config_name)
+
     project_path = args.project_path
     if not os.path.isdir(project_path):
         LOGGER.error(f"Invalid project path: {project_path}")
@@ -98,16 +135,21 @@ def main():
     trials = [
         f.split(".")[0] for f in os.listdir(project_path) if f.endswith(".Trial.enf")
     ]
+    included_trials = args.include.split(",") if args.include else []
+    if included_trials:
+        trials = [t for t in trials if t in included_trials]
+
     excluded_trials = args.exclude.split(",") if args.exclude else []
-    trials = [t for t in trials if t not in excluded_trials]
+    if excluded_trials:
+        trials = [t for t in trials if t not in excluded_trials]
 
     LOGGER.info(f"Trials: {', '.join(trials)}")
 
     pipeline_runner_path = os.path.join(
-        directory.SCRIPTS_DIR, "nushu_pipeline_runner", "nushu_pipeline_runner.py"
+        directory.SCRIPTS_DIR, "pipeline_runner", "pipeline_runner.py"
     )
 
-    common_args = ["-e", "-pp", project_path]
+    common_args = ["--export", "-cn", config_path, "-pn", project_path]
     if args.verbose:
         common_args.append("-v")
     if args.log:
@@ -124,7 +166,7 @@ def main():
         process = dist_utils.run_python_script(
             pipeline_runner_path,
             script_args,
-            # verbose=args.verbose,
+            verbose=args.verbose,
             python_path=sys.executable,
         )
         if process.returncode != 0:
@@ -138,6 +180,7 @@ def main():
     LOGGER.info(
         "Batch Runner complete. Time elapsed: {:.1f}s".format(time.time() - start)
     )
+    exit(0)
 
 
 if __name__ == "__main__":